Revert "Working up to ACPI reading and parsing, refactoring printf and klog, started on a kalloc routine"
This reverts commit 75e140471a228f8f893d1bac2a28fe90904727ad.
