--- .editorconfig
@@ -1,14 +1,0 @@
-root = true
-
-[*]
-end_of_line = lf
-insert_final_newline = true
-max_line_length = 120
-
-[*.{c,h}]
-charset = utf-8
-indent_style = space
-indent_size = 4
-# these might not be supported by all editors (are just a suggestion)
-curly_bracket_next_line = true
-spaces_around_operators = true
