--- TODO.md
@@ -1,7 +1,0 @@
-# General TODO list, in rough order of priority/timeline
-
-    It feels a bit high-fallutin to call this a "roadmap", so it's rather a loose collection of TODO items
-
-* Build network stack [in progress]
-* Build web server
-*
