+++ kernel/src/main.c
@@ -2,10 +2,7 @@
 #include <stddef.h>
 #include <limine.h>
 #include <term/term.h>
-#include <devicetree/dtb.h>
-#include <klog/klog.h>
-#include <net/network.h>
-#include <acpi/acpi.h>
+#include "devicetree/dtb.h"
 #include <serial/serial.h>
 
 
@@ -14,19 +11,71 @@
 // be made volatile or equivalent.
 
 static volatile struct limine_framebuffer_request framebuffer_request = {
-        .id = LIMINE_FRAMEBUFFER_REQUEST,
-        .revision = 0
+    .id = LIMINE_FRAMEBUFFER_REQUEST,
+    .revision = 0
 };
 
 static volatile struct limine_dtb_request dtb_request = {
-        .id = LIMINE_DTB_REQUEST,
-        .revision = 0,
+    .id = LIMINE_DTB_REQUEST,
+    .revision = 0,
 };
 
-static volatile struct limine_rsdp_request rsdp_request = {
-        .id = LIMINE_RSDP_REQUEST,
-        .revision = 0,
-};
+// GCC and Clang reserve the right to generate calls to the following
+// 4 functions even if they are not directly called.
+// Implement them as the C specification mandates.
+// DO NOT remove or rename these functions, or stuff will eventually break!
+// They CAN be moved to a different .c file.
+
+void *memcpy(void *dest, const void *src, size_t n) {
+    uint8_t *pdest = (uint8_t *)dest;
+    const uint8_t *psrc = (const uint8_t *)src;
+
+    for (size_t i = 0; i < n; i++) {
+        pdest[i] = psrc[i];
+    }
+
+    return dest;
+}
+
+void *memset(void *s, int c, size_t n) {
+    uint8_t *p = (uint8_t *)s;
+
+    for (size_t i = 0; i < n; i++) {
+        p[i] = (uint8_t)c;
+    }
+
+    return s;
+}
+
+void *memmove(void *dest, const void *src, size_t n) {
+    uint8_t *pdest = (uint8_t *)dest;
+    const uint8_t *psrc = (const uint8_t *)src;
+
+    if (src > dest) {
+        for (size_t i = 0; i < n; i++) {
+            pdest[i] = psrc[i];
+        }
+    } else if (src < dest) {
+        for (size_t i = n; i > 0; i--) {
+            pdest[i-1] = psrc[i-1];
+        }
+    }
+
+    return dest;
+}
+
+int memcmp(const void *s1, const void *s2, size_t n) {
+    const uint8_t *p1 = (const uint8_t *)s1;
+    const uint8_t *p2 = (const uint8_t *)s2;
+
+    for (size_t i = 0; i < n; i++) {
+        if (p1[i] != p2[i]) {
+            return p1[i] < p2[i] ? -1 : 1;
+        }
+    }
+
+    return 0;
+}
 
 // Halt and catch fire function.
 static void hcf(void) {
@@ -43,7 +92,7 @@ void _start(void) {
 
     // Ensure we got a framebuffer.
     if (framebuffer_request.response == NULL
-        || framebuffer_request.response->framebuffer_count < 1) {
+     || framebuffer_request.response->framebuffer_count < 1) {
         print_serial("ERR: No framebuffer\r\n");
         hcf();
     }
@@ -55,37 +104,24 @@ void _start(void) {
 
 
     term_init(
-            framebuffer->address,
-            framebuffer->width,
-            framebuffer->height,
-            framebuffer->pitch
-    );
+            framebuffer->address, framebuffer->width, framebuffer->height, framebuffer->pitch
+            );
 
     print_serial("Switching to tty print\r\n");
 
     TERM_WRITE_BUF("=== WebOS v0.0.1a (snapshot release) ===\n");
-    klog("main", "Loading DTB");
+    TERM_WRITE_BUF("Parsing DTB...");
     // Ensure we got a DTB
     if (dtb_request.response == NULL) {
-        klog("main", "no DTB present, wrong platform or booting in legacy mode?");
+        TERM_WRITE_BUF("NOPE\r\nDTB Not found (wrong platform, or not booting in UEFI mode?)");
     } else {
         dtb_parse(dtb_request.response->dtb_ptr);
-        klog("main", "DTB parsing complete");
+        TERM_WRITE_BUF("Done\n");
     }
 
-    klog("main", "Loading RSDP");
-    if (rsdp_request.response == NULL) {
-        klog("main", "no RSDP present");
-        hcf(); // panic
-    } else {
-        acpi_init(rsdp_request.response->address);
-    }
+    TERM_WRITE_BUF("Setting up network...");
 
-    klog("main", "Starting network driver");
-
-    net_init();
-
-    klog("main", "Network driver started");
+    TERM_WRITE_BUF("Done\n");
 
     // We're done, just hang...
     hcf();
