--- kernel/src/mem/kalloc.c
@@ -1,11 +1,0 @@
-#include "kalloc.h"
-
-void *kalloc(size_t len)
-{
-  
-}
-
-void kfree(void *ptr)
-{
-  
-}
