--- kernel/src/stdlib/README.md
@@ -1,5 +1,0 @@
-# STDLIB Stuff
-
-Do not include any headers in this directory.  The only functions you should expose
-from here are functions which are from the C standard library.  You can define more functions,
-but they should only be used in the file they are defined in.
