+++ kernel/src/term/term.c
@@ -3,13 +3,8 @@
 
 #include <flanterm/flanterm.h>
 #include <flanterm/backends/fb.h>
-#include <stdio.h>
-#include "klog/klog.h"
-#include "term.h"
 
 static struct flanterm_context* ctx;
-
-void term_putc(const char c);
 
 void term_init(uint32_t *framebuffer, size_t width, size_t height, size_t pitch)
 {
@@ -19,23 +14,4 @@ void term_init(uint32_t *framebuffer, size_t width, size_t height, size_t pitch)
 void term_write(const char* string, size_t count)
 {
     flanterm_write(ctx, string, count);
-}
-
-void term_putc(const char c)
-{
-    term_write(&c, 1);
-}
-
-// todo: move this to string.h/string.c
-// unsafe strlen function, counts until it hits a NULL byte with no limit enforced
-size_t strlen(const char* str)
-{
-    const char* s = str;
-    while(*(s++));
-    return s - str;
-}
-
-void term_printf(const char* fmt, ...)
-{
-
 }
